+++ functions/public.ps1
@@ -1,41 +1,132 @@
 
 #region public functions
 
-Function Get-mySQLiteTable {
+Function Convert-MySQLiteDB {
     [cmdletbinding()]
+    [alias('Export-DB')]
+    [outputtype("Object")]
+    Param(
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName)]
+        [ValidateNotNullOrEmpty()]
+        [alias("database", "fullname")]
+        [string]$Path,
+        [Parameter(Mandatory, HelpMessage = "Enter the name of the table with data to import")]
+        [ValidateNotNullOrEmpty()]
+        [string]$TableName
+    )
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+
+    } #begin
+
+    Process {
+        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Processing database $Path "
+        $file = resolvedb -Path $path
+        if ($file.exists) {
+            $connection = opendb -Path $file.path
+        }
+        else {
+            Throw "Failed to find database file $($file.path)"
+        }
+        #verify table exists
+        $tables = Get-mySQLiteTable -connection $connection -KeepAlive
+        if ($tables.name -contains $tablename) {
+            $query = "Select * from $tablename"
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Importing from table $Tablename"
+            $raw = Invoke-mySQLiteQuery -connection $connection -Query $query -As object -KeepAlive
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($raw.count) items"
+
+            #use propertymap table if it exists
+            if ($tables.name -contains "PropertyMap") {
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting property map"
+                $map = Invoke-mySQLiteQuery -Connection $connection -Query "Select * from propertymap" -KeepAlive -as Hashtable
+                foreach ($item in $raw) {
+                    $tmpHash = [ordered]@{}
+                    foreach ($key in $map.keys) {
+                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ...$key"
+                        $name = $key
+                        #if value of the raw object is byte[], assume it is an exported clixml file
+                        if ($item.$key.gettype().name -eq 'Byte[]') {
+                            $v = frombytes $item.$key
+                        }
+                        else {
+                            $v = $item.$key
+                        }
+                        $value = $v -as $($($map[$key] -as [type]))
+                        $tmpHash.Add($name,$value)
+                    }
+                    New-Object -typename PSObject -property $tmpHash
+                }
+            }
+            else {
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Writing raw objects to the pipeline"
+                $raw
+            }
+        }
+        else {
+            Write-Warning "Failed to find a table called $Tablename in $($file.path)"
+        }
+
+    } #process
+
+    End {
+        closedb $connection
+        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+
+    } #end
+
+} #close Convert-MySQLiteDB
+
+Function Get-mySQLiteTable {
+    [cmdletbinding(DefaultParameterSetName = "file")]
     [Alias("gtb", "Get-DBTable")]
     [outputtype("PSCustomObject")]
 
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.")]
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName, ParameterSetName = "file")]
         [ValidateNotNullOrEmpty()]
-        [alias("database")]
-        [string]$Path
+        [alias("database", "fullname")]
+        [string]$Path,
+        [Parameter(Position = 0, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
+        [System.Data.SQLite.SQLiteConnection]$Connection,
+        [Parameter(HelpMessage = "Do not close the connection.", ParameterSetName = "connection")]
+        [switch]$KeepAlive
     )
     Begin {
         Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting tables from $Path"
-        $db = resolvedb -Path $path
-        if ($db.exists) {
-            Invoke-mySQLiteQuery -Path $db.path -query "Select Name from sqlite_master where type='table'"
-        }
+        if ($pscmdlet.ParameterSetName -eq 'file') {
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting tables from $Path"
+            $db = resolvedb -Path $path
+            if ($db.exists) {
+                Invoke-mySQLiteQuery -Path $db.path -query "Select Name from sqlite_master where type='table'"
+            }
+            else {
+                Write-Warning "Failed to find $($db.path)"
+            }
+        } #if file parameter set
         else {
-            Write-Warning "Failed to find $($db.path)"
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using an existing connection $($connection.ConnectionString)"
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] KeepAlive is $KeepAlive"
+            Invoke-mySQLiteQuery -connection $Connection -query "Select Name from sqlite_master where type='table'" -KeepAlive:$KeepAlive
         }
     } #process
 
     End {
+        if ($connection -AND (-Not $KeepAlive)) {
+            closedb -connection $connection
+        }
+
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end
 
 } #close Get-mySQLiteTable
 
-Function Export-mySQLiteDB {
+Function ConvertTo-mySQLiteDB {
     [cmdletbinding(SupportsShouldProcess)]
-    [alias("todb",'Export-DB')]
+    [alias("todb", 'Convert-DB')]
     [outputtype("None")]
     Param(
         [Parameter(Mandatory, ValueFromPipeline, HelpMessage = "What object do you want to create")]
@@ -111,12 +202,12 @@ Function Export-mySQLiteDB {
                 $thash = [ordered]@{}
                 Foreach ($prop in $properties) {
                     Switch -Regex ($prop.TypeNameofValue) {
-                        "Int32$" { $sqltype = "Int" }
-                        "Int64$" {$sqltype = "Real"}
-                        "^System.Double$" { $sqltype = "Real"}
-                        "^System.DateTime" {$sqltype = "Text"}
-                        "^System.String$" {$sqltype = "Text"}
-                        "^System.Boolean$" {$sqltype = "Text"}
+                        "Int32$"            {$sqltype = "Int" }
+                        "Int64$"            {$sqltype = "Real"}
+                        "^System.Double$"   {$sqltype = "Real"}
+                        "^System.DateTime"  {$sqltype = "Text"}
+                        "^System.String$"   {$sqltype = "Text"}
+                        "^System.Boolean$"  {$sqltype = "Text"}
                         default {
                             $sqltype = "Blob"
                         }
@@ -144,7 +235,7 @@ Function Export-mySQLiteDB {
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end
 
-} #close Export-mySQLiteDB
+} #close Convert-mySQLiteDB
 
 Function New-mySQLiteDB {
     [cmdletbinding(SupportsShouldProcess)]
@@ -271,7 +362,7 @@ Function New-mySQLiteDBTable {
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Testing for table $Tablename"
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($cmd.CommandText)"
 
-                $ds = New-Object system.Data.DataSet
+                $ds = New-Object System.Data.DataSet
                 $da = New-Object System.Data.SQLite.SQLiteDataAdapter($cmd)
                 [void]$da.fill($ds)
 
@@ -301,7 +392,7 @@ Function New-mySQLiteDBTable {
                     $keys = $ColumnProperties.Keys
                     $primary = $keys | Select-Object -first 1`
                     $primaryType = $ColumnProperties.item($Primary)
-                    $cols = $keys | Select-object -skip 1
+                    $cols = $keys | Select-Object -skip 1
                     Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Setting Primary Key: $primary <$PrimaryType>"
                     $query += "($Primary $PrimaryType PRIMARY KEY"
                     foreach ($col in $cols) {
@@ -338,33 +429,45 @@ Function New-mySQLiteDBTable {
 } #new-mysqlitedbtable
 
 Function Invoke-mySQLiteQuery {
-    [cmdletbinding(SupportsShouldProcess)]
+    [cmdletbinding(SupportsShouldProcess, DefaultParameterSetName = "file")]
     [alias("iq")]
     [outputtype("None", "PSCustomObject", "System.Data.Datatable")]
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.")]
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ParameterSetName = "file", ValueFromPipelineByPropertyName)]
         [Alias("fullname", "database")]
         [ValidateNotNullOrEmpty()]
         [string]$Path,
+        [Parameter(Position = 0, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
+        [System.Data.SQLite.SQLiteConnection]$Connection,
         [Parameter(Position = 1, Mandatory, HelpMessage = "Enter a SQL query string")]
         [string]$Query,
-        [Parameter(HelpMessage = "Write the results of a Select query as a System.Data.Datatable")]
-        [switch]$AsDataTable
+        [Parameter(HelpMessage = "Keep the connection alive.",ParameterSetName = "connection")]
+        [switch]$KeepAlive,
+        [Parameter(HelpMessage = "Write the results of a Select query in the specified format")]
+        [ValidateSet("Object","Datatable","Hashtable")]
+        [string]$As = "object"
     )
     Begin {
         Write-Verbose "[$((Get-Date).TimeofDay) BEGIN] Starting $($myinvocation.mycommand)"
-        #resolve or convert path into a full filesystem path
-        $path = $executioncontext.sessionstate.path.GetUnresolvedProviderPathFromPSPath($path)
-
-        If (Test-Path -Path $path) {
-            $connection = opendb -Path $path
-        }
-        else {
-            Write-Warning "Cannot find the database file $path."
-        }
 
     } #begin
     Process {
+        if ($pscmdlet.ParameterSetName -eq 'file') {
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using file $path"
+            $file = resolvedb -path $path
+
+            If ($file.exists) {
+                $connection = opendb -Path $path
+            }
+            else {
+                Write-Warning "Cannot find the database file $path."
+            }
+        }
+        else {
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using existing connection $($connection.ConnectionString)"
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] KeepAlive is $KeepAlive"
+        }
+
         Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $query"
         if ($connection.state -eq 'Open' -OR $PSBoundparameters.ContainsKey("WhatIf")) {
 
@@ -375,17 +478,16 @@ Function Invoke-mySQLiteQuery {
                 #determine what method to invoke based on the query
                 Switch -regex ($query) {
                     "^Select (\w+|\*)|(@@\w+ AS)" {
-                        if ($AsDataTable) {
+                        if ($As -eq "datatable") {
                             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Datatable"
                             $ds = New-Object System.Data.DataSet
                             $da = New-Object System.Data.SQLite.SQLiteDataAdapter($cmd)
-                            $da.fill($ds)
+                            [void]$da.fill($ds)
                             $ds.Tables
                         }
                         else {
                             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ExecuteReader"
                             $reader = $cmd.executereader()
-                            $out = @()
                             #convert datarows to a custom object
                             while ($reader.read()) {
 
@@ -395,10 +497,16 @@ Function Invoke-mySQLiteQuery {
 
                                     $h.add($col, $reader.getvalue($i))
                                 } #for
-                                $out += New-Object -TypeName psobject -Property $h
+
+                                if ($as -eq "hashtable") {
+                                    $h
+                                }
+                                else {
+                                    New-Object -TypeName psobject -Property $h
+                                }
+
                             } #while
 
-                            $out
                             $reader.close()
                         }
                         Break
@@ -422,11 +530,10 @@ Function Invoke-mySQLiteQuery {
         }
     } #process
     End {
-        if ($connection.state -eq 'Open') {
+        #if the connection was passed as a parameter, do not close it. The generating command is responsible
+        if ( (($connection.state -eq 'Open') -AND ($pscmdlet.ParameterSetName -eq 'file')) -OR (($connection.state -eq 'Open') -AND (-Not $KeepAlive)) ) {
             Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
-            $cmd.Dispose()
-            $connection.close()
-            $connection.Dispose()
+            closedb -connection $connection -cmd $cmd
         }
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end