+++ mySQLite.psm1
@@ -1,10 +1,8 @@
 
-#todo: add Append option to Export-mySQLiteDB
-#todo: create an Import-mySQLiteDB command
 
 Add-Type -Path $PSScriptRoot\assembly\system.data.sqlite.dll
 
-get-childitem $psscriptroot\functions\*.ps1 |
+Get-ChildItem -path $psscriptroot\functions\*.ps1 |
 Foreach-Object {
     . $_.fullname
 }